Code reorg, typos
Previously, the main switch in parseStatement() would compare starttype to Objective-J keywords even if Objective-J support was off. Then each of the Objective-J statement cases checked options.objj individually, which resulted in the somewhat confusing construction of a return followed by a break at the end of each case.

Now the Objective-J statements are handled separately in the default case of the main switch, and options.objj only has to be checked once, which has the added benefit of eliminating the comparisons to Objective-J keywords when options.objj is false.